requirements.txt
pandas==2.2.2
streamlit==1.37.1
plotly-express==0.4.1